Rename package gatherinfocontract to gatherinfo
