+++ repairnator/repairnator-pipeline/src/main/java/fr/inria/spirals/repairnator/process/step/ComputeSourceDir.java
@@ -3,7 +3,6 @@ package fr.inria.spirals.repairnator.process.step;
 import fr.inria.spirals.repairnator.ProjectState;
 import fr.inria.spirals.repairnator.process.inspectors.ProjectInspector;
 import fr.inria.spirals.repairnator.process.maven.MavenHelper;
-import fr.inria.spirals.repairnator.process.step.gatherinfocontract.GatherTestInformation;
 import org.apache.maven.model.Build;
 import org.apache.maven.model.Model;
 import org.apache.maven.model.building.ModelBuildingException;
