+++ repairnator/repairnator-pipeline/src/main/java/fr/inria/spirals/repairnator/process/step/NopolRepair.java
@@ -11,7 +11,6 @@ import fr.inria.spirals.repairnator.process.inspectors.ProjectInspector;
 import fr.inria.spirals.repairnator.process.nopol.IgnoreStatus;
 import fr.inria.spirals.repairnator.process.nopol.NopolInformation;
 import fr.inria.spirals.repairnator.process.nopol.NopolStatus;
-import fr.inria.spirals.repairnator.process.step.gatherinfocontract.GatherTestInformation;
 import fr.inria.spirals.repairnator.process.testinformation.ComparatorFailureLocation;
 import fr.inria.spirals.repairnator.process.testinformation.FailureLocation;
